Add istio ingress rules
By default istio blocks all inbound traffic to hosts outside the
cluster. Therefore we must explicitly specify a route into the
blunderlist service for the outside world to access. This is achieved
with an ingress gateway and virtual service.

Separate virtual services are defined to serve both the frontend and
gateway services respectively using the URI path prefix to route
requests to the backends they are destined.